+++ CHANGELOG.md
@@ -3,6 +3,7 @@
   - [#174](https://github.com/Datatamer/unify-client-python/issues/174) Get and create taxonomy categories
   - [#182](https://github.com/Datatamer/unify-client-python/issues/182) Add the ability to refresh estimated pair counts.
   - [#184](https://github.com/Datatamer/unify-client-python/issues/184) Support for getting published cluster configurations
+  - [#201](https://github.com/Datatamer/unify-client-python/issues/201) Support for refreshing published cluster IDs
 
 ## 0.7.0
   **BREAKING CHANGES**